pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vllmd"
version = "0.1.0"
description = "Video action recognition with multimodal support for early identification of substance-related risk behaviors in school environments."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = []
keywords = ["video", "action-recognition", "llm", "pytorch"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "numpy",
    "opencv-python",
    "pandas",
    "scikit-learn",
    "torch",
    "torchvision",
    "matplotlib",
    "requests",
    "Pillow",
    "python-dotenv",
]

[project.optional-dependencies]
gemini = ["google-genai"]
dev = ["pytest", "ruff", "pyright"]
transformers = ["transformers"]

[project.scripts]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-p no:cacheprovider"

[tool.pyright]
extraPaths = ["src"]
